requirements.txt
MySQL-python
matplotlib
lxml
python-docx
